+++ code/src/cljs/sixsq/nuvla/ui/apps_application/views.cljs
@@ -115,12 +115,60 @@
        :default-open false])))
 
 
+(defn DockerComposeValidationPopup
+  [{:keys [loading? valid? error-msg] :as validate-dc}]
+  [ui/Popup
+   {:trigger        (r/as-element [ui/Icon {:name    (cond
+                                                       loading? "circle notched"
+                                                       (not valid?) "bug"
+                                                       valid? "checkmark")
+                                            :color   (cond
+                                                       loading? "black"
+                                                       (not valid?) "red"
+                                                       valid? "green")
+                                            :loading loading?}])
+    :header         "Validation of docker-compose"
+    :content        (cond
+                      loading? "Validation of docker-compose in progress..."
+                      (not valid?) (some-> error-msg (str/replace #"^.*is invalid because:" ""))
+                      valid? "Docker-compose is valid :)")
+    :on             "hover"
+    :position       "top center"
+    :wide           true
+    :hide-on-scroll true}])
+
+
+(defn DockerComposeCompatibility
+  [compatibility unsupp-opts]
+  (let [popup-disabled? (empty? unsupp-opts)]
+    [:div {:style {:float "right"}}
+     [:span {:style {:font-variant "small-caps"}} "compatibility: "]
+     [ui/Label {:color      "blue"
+                :horizontal true} compatibility]
+     (when-not popup-disabled?
+       [ui/Popup
+        {:trigger        (r/as-element [ui/Icon {:color "yellow"
+                                                 :name  "exclamation triangle"}])
+         :header         "Unsupported options"
+         :content        (str "Swarm doesn't support and will ignore the following options: "
+                              (str/join "; " unsupp-opts))
+
+         :on             "hover"
+         :position       "top right"
+         :wide           true
+         :hide-on-scroll true}])]))
+
+
 (defn docker-compose-section []
   (let [tr              (subscribe [::i18n-subs/tr])
         docker-compose  (subscribe [::subs/docker-compose])
+        module-app      (subscribe [::apps-subs/module])
+        unsupp-opts     (:unsupported-options (:content @module-app))
+        compatibility   (:compatibility @module-app)
         form-valid?     (subscribe [::apps-subs/form-valid?])
         editable?       (subscribe [::apps-subs/editable?])
         module-subtype  (subscribe [::apps-subs/module-subtype])
+        validate-dc     (subscribe [::apps-subs/validate-docker-compose])
         local-validate? (r/atom false)
         default-value   @docker-compose]
     (fn []
@@ -128,7 +176,9 @@
         [uix/Accordion
          [:<>
           [:div {:style {:margin-bottom "10px"}} "Env substitution"
-           [:span ff/nbsp (ff/help-popup (@tr [:module-docker-compose-help]))]]
+           [:span ff/nbsp (ff/help-popup (@tr [:module-docker-compose-help]))]
+           (when (= @module-subtype "application")
+             [DockerComposeCompatibility compatibility unsupp-opts])]
 
           [ui/CodeMirror {:value      default-value
                           :autoCursor true
@@ -147,7 +197,11 @@
                (if (str/blank? error-msg)
                  (@tr [:module-docker-compose-error])
                  error-msg)]))]
-         :label (if (= @module-subtype "application") "Docker compose" "Manifest")
+         :label (if (= @module-subtype "application")
+                  [:span "Docker compose" ff/nbsp
+                   (when @validate-dc
+                     [DockerComposeValidationPopup @validate-dc])]
+                  "Manifest")
          :default-open true]))))